refactor: move tables to component/table
